Restructure modules for better cohesion.
