+++ app/app.py
@@ -5,8 +5,8 @@ from flask import Flask
 
 from app import jwt_handlers
 from app.extensions import database, jwt
+from app.routes.applicant.personal_info_routes import personal_info_bp
 from app.routes.competences_routes import competences_bp
-from app.routes.personal_info_routes import personal_info_bp
 
 
 def create_app() -> Flask:
@@ -67,9 +67,10 @@ def register_blueprints(application_form_api: Flask) -> None:
     """
 
     application_form_api.register_blueprint(
-            personal_info_bp, url_prefix='/applicant/personal_info')
+            personal_info_bp,
+            url_prefix='/application-form/applicant/personal_info')
     application_form_api.register_blueprint(
-            competences_bp, url_prefix='/competences')
+            competences_bp, url_prefix='/application-form/competences')
 
 
 if __name__ == "__main__":
