+++ tests/routes/applicant/test_personal_info_routes.py
@@ -13,7 +13,7 @@ def test_get_personal_info_success(app_with_client):
     token = generate_token_for_user_id_1(app)
 
     response = test_client.get(
-            '/applicant/personal_info/',
+            '/application-form/applicant/personal_info/',
             headers={'Authorization': f'Bearer {token}'})
     assert response.status_code == 200
     assert response.json['name'] == 'test'
@@ -27,7 +27,7 @@ def test_get_personal_info_no_result(app_with_client):
     token = generate_token_for_user_id_1(app)
 
     response = test_client.get(
-            '/applicant/personal_info/',
+            '/application-form/applicant/personal_info/',
             headers={'Authorization': f'Bearer {token}'})
     assert response.status_code == 404
     assert response.json['error'] == 'USER_NOT_FOUND'
@@ -38,12 +38,12 @@ def test_get_personal_info_sqlalchemy_error(app_with_client):
     app, test_client = app_with_client
     token = generate_token_for_user_id_1(app)
 
-    with patch('app.routes.personal_info_routes.'
+    with patch('app.routes.applicant.personal_info_routes.'
                'fetch_personal_info') as mock_fetch:
         mock_fetch.side_effect = SQLAlchemyError("A database error occurred")
 
         response = test_client.get(
-                '/applicant/personal_info/',
+                '/application-form/applicant/personal_info/',
                 headers={'Authorization': f'Bearer {token}'})
 
         assert response.status_code == 500
