+++ tests/test_jwt_handlers.py
@@ -13,8 +13,9 @@ def test_valid_token(app_with_client):
         valid_token = create_access_token(
                 identity={'id': 1}, expires_delta=datetime.timedelta(days=1))
 
-    response = test_client.get('/applicant/personal_info/', headers={
-        'Authorization': f'Bearer {valid_token}'})
+    response = test_client.get('/application-form/applicant/personal_info/',
+                               headers={
+                                   'Authorization': f'Bearer {valid_token}'})
     assert response.status_code == 200
 
 
@@ -26,8 +27,9 @@ def test_invalid_token(app_with_client):
                 identity={'id': 1}, expires_delta=datetime.timedelta(days=1))
         invalid_token = valid_token + 'invalid'
 
-    response = test_client.get('/applicant/personal_info/', headers={
-        'Authorization': f'Bearer {invalid_token}'})
+    response = test_client.get('/application-form/applicant/personal_info/',
+                               headers={
+                                   'Authorization': f'Bearer {invalid_token}'})
     assert response.status_code == 401
     assert response.json['error'] == 'INVALID_TOKEN'
 
@@ -40,14 +42,15 @@ def test_expired_token(app_with_client):
         expired_token = create_access_token(
                 identity={'id': 1}, expires_delta=datetime.timedelta(days=-1))
 
-    response = test_client.get('/applicant/personal_info/', headers={
-        'Authorization': f'Bearer {expired_token}'})
+    response = test_client.get('/application-form/applicant/personal_info/',
+                               headers={
+                                   'Authorization': f'Bearer {expired_token}'})
     assert response.status_code == 401
     assert response.json['error'] == 'TOKEN_EXPIRED'
 
 
 def test_unauthorized_request(app_with_client):
     app, test_client = app_with_client
-    response = test_client.get('/applicant/personal_info/')
+    response = test_client.get('/application-form/applicant/personal_info/')
     assert response.status_code == 401
     assert response.json['error'] == 'UNAUTHORIZED'
